allow only owner can acess their video's delete and edit

+++ src/controllers/videoController.js
@@ -8,6 +8,9 @@ export const home = async (req, res) => {
 
 export const see = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id).populate("owner");
   if (video) {
     return res.render("watch", { pageTitle: video.title, video });
@@ -17,25 +20,37 @@ export const see = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
-  if (video) {
-    return res.render("edit", { pageTitle: `${video.title}`, video });
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found" });
   }
-  return res.status(404).render("404", { pageTitle: "Video not found" });
+  if (String(_id) !== String(video.owner)) {
+    return res.status(403).redirect("/");
+  }
+  return res.render("edit", { pageTitle: `${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
-  if (video) {
-    await Video.findByIdAndUpdate(id, {
-      title,
-      description,
-      hashtags: Video.formatHashtags(hashtags),
-    });
+  if (!video) {
+    return res.render("404", { pageTitle: "Video not found" });
   }
-  return res.render("404", { pageTitle: "Video not found" });
+  if (String(_id) !== String(video.owner)) {
+    return res.status(403).redirect("/");
+  }
+  await Video.findByIdAndUpdate(id, {
+    title,
+    description,
+    hashtags: Video.formatHashtags(hashtags),
+  });
 };
 
 export const remove = (req, res) => {
@@ -76,6 +91,16 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+  if (String(_id) !== String(video.owner)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
